Improved the code and separated login and registration into components
